Add Compare link to README.md
I added a link to compare changes in `main` vs upstream `master`, which shows no changes. You can select another head branch from the `/compare` menu, and feel free to make any changes.

So I'm a little confused. Is there code? Was there a PR that made this unnecessary? Please update the docs with the latest status, in the interest of transparency. Thanks!