Simplify evil_move_word_impl, fixing bugs
The function `evil_move_word_impl` can be simplified, thereby fixing the
following two bugs in the process:

1. In select mode, the key sequence `wb` would result in a selection,
   instead of being a no-op as in Vim.

2. The functions `evil_next_word_end` and `evil_next_long_word_end`
   (mapped to 'e' and 'E') in normal mode would make the cursor land on
   the space after the word, instead of the last character of the word.

The only problem with this commit is that now the functions
`evil_next_word_start` and `evil_next_long_word_start` (mapped to `w`
and `W`) don't behave correctly anymore, making the cursor land
on the space before the next word. We'll fix it in the next commit.